requirements.txt
python-dotenv==0.19.2
python-telegram-bot==13.11
requests==2.27.0
